+++ packages/grid/src/core/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './_id-creator'
+export * from './id-creator'
 export * from './field'
 export * from './common'
 export * from './weak-map'
