+++ src/app/component/table.component.html
@@ -6,7 +6,7 @@
 </div>
 <thy-tabs (thyActiveTabChange)="activeTabChange($event)" [thyActiveTab]="tableService.activeView()._id">
     @for (item of tableService.views(); track $index) {
-        <thy-tab [thyTitle]="item.text" [_id]="item._id">
+        <thy-tab [thyTitle]="item.text" [id]="item._id">
             @if (item.isActive) {
                 <router-outlet></router-outlet>
             }
